Change renderOrder and depthWrite properties of all materials so that transparent line do not occlude visible ones.
See
https://stackoverflow.com/questions/15994944/transparent-objects-in-threejs
for more details. Using depthWrite = true on the gray lines is much
prettier and gives a nice sense of depth (go figure). But we when the
opacity is zero, we need depthWrite = false to solve the issues brought
up in the above referenced stackoverflow question. So we use an if
statement inside the opacity control box to change depthWrite.