Draw subject lines and mean lines separately in tract-details.js.
This commit reduces redundant line drawing in tract-details.js. Previous
versions appended the mean lines to trLines and asked d3 to update the
svg line elements for all of the data in trLines. While d3 is good about
reducing work using the enter(), update(), and exit() sets, we already
know exactly which lines will be updated based on the user input. So we
can avoid checking all of the lines in trLines entirely. The subject
lines are updated only when the user updates the plot type (e.g. fa or
md). Otherwise, only the mean lines are updated. On Adam's 2013 macbook
pro this reduced the update-on-sort times from approximately 1.5 seconds
to 0.6 seconds.